Commit
Commit of project
TODO - In the year since I wrote this, there seems to be some slight changes to how vpython works in jupyter. When I last left off, this project allowed me to switch between traversing the system in cartesian, spherical polar, or cylindrical coordinates. However it is now stuck such that the camera cannot be moved, making the program all but useless